moved interface tests around and moved PHOENIX model download into a method in grid_tools/interfaces.py
